requirements.txt
Babel==2.6.0
beautifulsoup4==4.6.3
Django==2.1.3
django-ckeditor==5.6.1
django-js-asset==1.1.0
django-phonenumber-field==2.1.0
django-resized==0.3.9
django-tinymce==2.7.0
mysqlclient==1.3.13
phonenumbers==8.10.0
Pillow==5.3.0
pytz==2018.7
sorl-thumbnail==12.5.0
